VIDCS-2956: Mic Panel Gets Displaced upon Resize (and some light refactoring) (#132)
