--- frontend/src/components/MeetingRoom/AudioControlButton/index.tsx
@@ -1,3 +1,0 @@
-import AudioControlButton from './AudioControlButton';
-
-export default AudioControlButton;
