--- frontend/src/components/MeetingRoom/AudioInputOutputDevices/index.tsx
@@ -1,3 +1,0 @@
-import AudioInputOutputDevices from './AudioInputOutputDevices';
-
-export default AudioInputOutputDevices;
