+++ frontend/src/components/MeetingRoom/DeviceControlButton/index.tsx
@@ -1,0 +1,3 @@
+import DeviceControlButton from './DeviceControlButton';
+
+export default DeviceControlButton;
