+++ frontend/src/components/MeetingRoom/DeviceSettingsMenu/index.tsx
@@ -1,0 +1,3 @@
+import DeviceSettingsMenu from './DeviceSettingsMenu';
+
+export default DeviceSettingsMenu;
