+++ frontend/src/components/MeetingRoom/ReduceNoiseTestSpeakers/index.tsx
