--- frontend/src/components/MeetingRoom/VideoControlButton/index.tsx
@@ -1,3 +1,0 @@
-import VideoControlButton from './VideoControlButton';
-
-export default VideoControlButton;
