--- frontend/src/components/MeetingRoom/VideoOutputDevices/index.tsx
@@ -1,3 +1,0 @@
-import VideoOutputDevices from './VideoOutputDevices';
-
-export default VideoOutputDevices;
